Crate ascii_canvas

An “ASCII Canvas” allows us to draw lines and write text into a fixed-sized canvas and then convert that canvas into ASCII characters. ANSI styling is supported.

Modules

style

The Style type is a simplified view of the various attributes offered by the term library. These are enumerated as bits so they can be easily or’d together etc.

Structs

AsciiCanvas

Row

ShiftedView

Gives a view onto an AsciiCanvas that has a fixed upper-left point. You can get one of these by calling the shift() method on any ASCII view.

StyleView

Gives a view onto an AsciiCanvas that applies an additional style to things that are written. You can get one of these by calling the styled() method on any ASCII view.

Traits

AsciiView

AsciiView is a view onto an AsciiCanvas which potentially applies transformations along the way (e.g., shifting, adding styling information). Most of the main drawing methods for AsciiCanvas are defined as inherent methods on an AsciiView trait object.