Module ascii_canvas::style

The Style type is a simplified view of the various attributes offered by the term library. These are enumerated as bits so they can be easily or’d together etc.

Structs

Style

StyleCursor

Constants

BG_BLACK

BG_BLUE

BG_BRIGHT_BLACK

BG_BRIGHT_BLUE

BG_BRIGHT_CYAN

BG_BRIGHT_GREEN

BG_BRIGHT_MAGENTA

BG_BRIGHT_RED

BG_BRIGHT_WHITE

BG_BRIGHT_YELLOW

BG_CYAN

BG_GREEN

BG_MAGENTA

BG_RED

BG_WHITE

BG_YELLOW

BLINK

BOLD

DEFAULT

DIM

FG_BLACK

FG_BLUE

FG_BRIGHT_BLACK

FG_BRIGHT_BLUE

FG_BRIGHT_CYAN

FG_BRIGHT_GREEN

FG_BRIGHT_MAGENTA

FG_BRIGHT_RED

FG_BRIGHT_WHITE

FG_BRIGHT_YELLOW

FG_CYAN

FG_GREEN

FG_MAGENTA

FG_RED

FG_WHITE

FG_YELLOW

ITALIC

REVERSE

SECURE

STANDOUT

UNDERLINE